refactor: separate Overpass and installation services
