+++ scripts/package-boundaries.json
@@ -10,27 +10,47 @@
       "src/data/localGeojsonCore.js",
       "src/data/localGeojsonLod.js"
     ],
-    "external": ["cesium"]
+    "external": [
+      "cesium"
+    ]
   },
   "application": {
-    "exports": ["./application"],
-    "modules": ["src/app/application.js"],
+    "exports": [
+      "./application"
+    ],
+    "modules": [
+      "src/app/application.js"
+    ],
     "external": []
   },
   "viewer": {
-    "exports": ["./application/viewer"],
-    "modules": ["src/app/viewer.js"],
-    "external": ["cesium"]
+    "exports": [
+      "./application/viewer"
+    ],
+    "modules": [
+      "src/app/viewer.js"
+    ],
+    "external": [
+      "cesium"
+    ]
   },
   "vite-build": {
     "runtime": "node",
-    "exports": ["./build/vite"],
-    "modules": ["build/vite.js"],
-    "external": ["vite-plugin-cesium"]
+    "exports": [
+      "./build/vite"
+    ],
+    "modules": [
+      "build/vite.js"
+    ],
+    "external": [
+      "vite-plugin-cesium"
+    ]
   },
   "live-providers": {
     "runtime": "node",
-    "exports": ["./server/providers/live"],
+    "exports": [
+      "./server/providers/live"
+    ],
     "modules": [
       "server/providers/live.js",
       "server/providers/common/http.js",
@@ -45,16 +65,24 @@
       "src/data/aisStreamAdapter.js",
       "src/data/aisWatchdog.js"
     ],
-    "external": ["ws"]
+    "external": [
+      "ws"
+    ]
   },
   "adsb-lol-source": {
-    "exports": ["./sources/adsb-lol"],
-    "modules": ["src/data/adsbLolFallback.js"],
+    "exports": [
+      "./sources/adsb-lol"
+    ],
+    "modules": [
+      "src/data/adsbLolFallback.js"
+    ],
     "external": []
   },
   "place-providers": {
     "runtime": "node",
-    "exports": ["./server/providers/places"],
+    "exports": [
+      "./server/providers/places"
+    ],
     "modules": [
       "server/providers/places.js",
       "server/providers/places/google.js",
@@ -69,13 +97,19 @@
     "external": []
   },
   "place-payloads": {
-    "exports": ["./sources/places"],
-    "modules": ["src/data/placeProviderPayloads.js"],
+    "exports": [
+      "./sources/places"
+    ],
+    "modules": [
+      "src/data/placeProviderPayloads.js"
+    ],
     "external": []
   },
   "space-providers": {
     "runtime": "node",
-    "exports": ["./server/providers/space"],
+    "exports": [
+      "./server/providers/space"
+    ],
     "modules": [
       "server/providers/space.js",
       "server/providers/space/celestrak.js",
@@ -86,13 +120,19 @@
     "external": []
   },
   "space-requests": {
-    "exports": ["./sources/space"],
-    "modules": ["src/data/spaceProviderRequests.js"],
+    "exports": [
+      "./sources/space"
+    ],
+    "modules": [
+      "src/data/spaceProviderRequests.js"
+    ],
     "external": []
   },
   "terrain-provider": {
     "runtime": "node",
-    "exports": ["./server/providers/terrain"],
+    "exports": [
+      "./server/providers/terrain"
+    ],
     "modules": [
       "server/providers/terrain.js",
       "src/data/terrainHeightsProxy.js"
@@ -100,41 +140,70 @@
     "external": []
   },
   "terrain-source": {
-    "exports": ["./sources/terrain"],
-    "modules": ["src/data/terrainHeightsProxy.js"],
+    "exports": [
+      "./sources/terrain"
+    ],
+    "modules": [
+      "src/data/terrainHeightsProxy.js"
+    ],
     "external": []
   },
   "traffic-provider": {
     "runtime": "node",
-    "exports": ["./server/providers/traffic"],
-    "modules": ["server/providers/traffic.js", "src/data/tomtomTiles.js"],
+    "exports": [
+      "./server/providers/traffic"
+    ],
+    "modules": [
+      "server/providers/traffic.js",
+      "src/data/tomtomTiles.js"
+    ],
     "external": []
   },
   "traffic-source": {
-    "exports": ["./sources/traffic"],
-    "modules": ["src/data/tomtomTiles.js"],
+    "exports": [
+      "./sources/traffic"
+    ],
+    "modules": [
+      "src/data/tomtomTiles.js"
+    ],
     "external": []
   },
   "firms-provider": {
     "runtime": "node",
-    "exports": ["./server/providers/firms"],
-    "modules": ["server/providers/firms.js", "src/data/firmsCsv.js"],
+    "exports": [
+      "./server/providers/firms"
+    ],
+    "modules": [
+      "server/providers/firms.js",
+      "src/data/firmsCsv.js"
+    ],
     "external": []
   },
   "gbfs-provider": {
     "runtime": "node",
-    "exports": ["./server/providers/gbfs"],
-    "modules": ["server/providers/gbfs.js", "src/data/gbfsSource.js"],
+    "exports": [
+      "./server/providers/gbfs"
+    ],
+    "modules": [
+      "server/providers/gbfs.js",
+      "src/data/gbfsSource.js"
+    ],
     "external": []
   },
   "gbfs-source": {
-    "exports": ["./sources/gbfs"],
-    "modules": ["src/data/gbfsSource.js"],
+    "exports": [
+      "./sources/gbfs"
+    ],
+    "modules": [
+      "src/data/gbfsSource.js"
+    ],
     "external": []
   },
   "cctv-provider": {
     "runtime": "node",
-    "exports": ["./server/providers/cctv"],
+    "exports": [
+      "./server/providers/cctv"
+    ],
     "modules": [
       "server/providers/cctv.js",
       "server/providers/cctv/catalog.js",
@@ -151,7 +220,9 @@
   },
   "radio-provider": {
     "runtime": "node",
-    "exports": ["./server/providers/radio"],
+    "exports": [
+      "./server/providers/radio"
+    ],
     "modules": [
       "server/providers/radio.js",
       "server/providers/radio/catalog.js",
@@ -160,6 +231,46 @@
       "server/providers/radio/transport.js",
       "src/data/radioCountry.js",
       "server/providers/common/http.js"
+    ],
+    "external": []
+  },
+  "overpass-provider": {
+    "runtime": "node",
+    "exports": [
+      "./server/providers/overpass"
+    ],
+    "modules": [
+      "server/providers/common/geo.js",
+      "server/providers/common/http.js",
+      "server/providers/common/rate-limit.js",
+      "server/providers/common/request.js",
+      "server/providers/overpass.js",
+      "server/providers/overpass/cache.js",
+      "server/providers/overpass/constants.js",
+      "server/providers/overpass/geometry.js",
+      "server/providers/overpass/query.js",
+      "server/providers/overpass/transport.js",
+      "server/providers/places/routes.js",
+      "src/data/placeProviderPayloads.js"
+    ],
+    "external": []
+  },
+  "military-installations-provider": {
+    "runtime": "node",
+    "exports": [
+      "./server/providers/military-installations"
+    ],
+    "modules": [
+      "server/providers/common/http.js",
+      "server/providers/common/query.js",
+      "server/providers/common/rate-limit.js",
+      "server/providers/military-installations.js",
+      "server/providers/military-installations/cache.js",
+      "server/providers/military-installations/constants.js",
+      "server/providers/military-installations/query.js",
+      "server/providers/overpass/constants.js",
+      "server/providers/overpass/geometry.js",
+      "server/providers/overpass/transport.js"
     ],
     "external": []
   }
